chore: Update package.json and pricing.ts
Update package.json to version 1.3.6 and add supportModelType 'gpt-4o-2024-05-13' in pricing.ts
