+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "gpt-tokens",
-    "version": "1.3.4",
+    "version": "1.3.6",
     "description": "Calculate the token consumption and amount of openai gpt message",
     "keywords": [
         "gpt",
