+++ src/index.ts
@@ -1,8 +1,18 @@
-import { encodingForModel, getEncoding, Tiktoken } from 'js-tiktoken'
-import { promptTokensEstimate }                    from 'openai-chat-tokens'
-import { TokenPrice }                              from './tokenPrice'
+import { encodingForModel, getEncoding } from 'js-tiktoken'
+import { promptTokensEstimate }          from 'openai-chat-tokens'
+import { TokenPrice }                    from './tokenPrice'
+import type { Tiktoken }                 from 'js-tiktoken'
+import type { supportModelType }         from './pricing'
 
-let modelEncodingCache: { [key in supportModelType]?: Tiktoken } = {}
+export type { supportModelType }
+
+interface MessageItem {
+    name?  : string
+    role   : 'system' | 'user' | 'assistant'
+    content: string
+}
+
+const modelEncodingCache: { [key in supportModelType]?: Tiktoken } = {}
 
 export function getEncodingForModelCached (model: supportModelType): Tiktoken {
     if (!modelEncodingCache[model]) {
@@ -18,48 +28,14 @@ export function getEncodingForModelCached (model: supportModelType): Tiktoken {
     return modelEncodingCache[model]!
 }
 
-/**
- * This is a port of the Python code from
- *
- * https://notebooks.githubusercontent.com/view/ipynb?browser=edge&bypass_fastly=true&color_mode=dark&commit=d67c4181abe9dfd871d382930bb778b7014edc66&device=unknown_device&docs_host=https%3A%2F%2Fdocs.github.com&enc_url=68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f6f70656e61692f6f70656e61692d636f6f6b626f6f6b2f643637633431383161626539646664383731643338323933306262373738623730313465646336362f6578616d706c65732f486f775f746f5f636f756e745f746f6b656e735f776974685f74696b746f6b656e2e6970796e62&logged_in=true&nwo=openai%2Fopenai-cookbook&path=examples%2FHow_to_count_tokens_with_tiktoken.ipynb&platform=mac&repository_id=468576060&repository_type=Repository&version=114#6d8d98eb-e018-4e1f-8c9e-19b152a97aaf
- */
-
-export type supportModelType =
-    | 'gpt-3.5-turbo'
-    | 'gpt-3.5-turbo-16k'
-    | 'gpt-4'
-    | 'gpt-4-32k'
-    | 'gpt-4-turbo-preview'
-    | 'gpt-3.5-turbo-0301'
-    | 'gpt-3.5-turbo-0613'
-    | 'gpt-3.5-turbo-1106'
-    | 'gpt-3.5-turbo-0125'
-    | 'gpt-3.5-turbo-16k-0613'
-    | 'gpt-4-0314'
-    | 'gpt-4-0613'
-    | 'gpt-4-32k-0314'
-    | 'gpt-4-32k-0613'
-    | 'gpt-4-1106-preview'
-    | 'gpt-4-0125-preview'
-    | 'gpt-4-turbo-2024-04-09'
-    | 'gpt-4-turbo'
-    | 'gpt-4o'
-    | 'gpt-4o-2024-05-13'
-
-interface MessageItem {
-    name?: string
-    role: 'system' | 'user' | 'assistant'
-    content: string
-}
-
 export class GPTTokens extends TokenPrice {
     constructor (options: {
-        model?: supportModelType
+        model?        : supportModelType
         fineTuneModel?: string
-        messages?: GPTTokens['messages']
-        training?: GPTTokens['training']
-        tools?: GPTTokens['tools']
-        debug?: boolean
+        messages?     : GPTTokens['messages']
+        training?     : GPTTokens['training']
+        tools?        : GPTTokens['tools']
+        debug?        : boolean
     }) {
         super()
 
@@ -118,16 +94,16 @@ export class GPTTokens extends TokenPrice {
     public readonly tools?: {
         type: 'function'
         function: {
-            name: string
+            name        : string
             description?: string
-            parameters: Record<string, unknown>
+            parameters  : Record<string, unknown>
         }
     } []
 
     // Used USD
     public get usedUSD () {
         if (this.training) return this.trainPrice(this.model, this.usedTokens)
-        if (this.tools) return this.inputPrice(this.model, this.usedTokens)
+        if (this.tools)    return this.inputPrice(this.model, this.usedTokens)
 
         return this.totalPrice(this.fineTuneModel
                 ? `ft:${this.model}`
@@ -207,9 +183,10 @@ export class GPTTokens extends TokenPrice {
      * @throws If the model is not supported.
      */
     private static num_tokens_from_messages (messages: MessageItem [], model: supportModelType) {
-        let encoding!: Tiktoken
+        let encoding!          : Tiktoken
         let tokens_per_message!: number
-        let tokens_per_name !: number
+        let tokens_per_name !  : number
+
         let num_tokens = 0
 
         if (model === 'gpt-3.5-turbo-0301') {
@@ -222,7 +199,12 @@ export class GPTTokens extends TokenPrice {
 
         encoding = getEncodingForModelCached(model)
 
-        // Python 2 Typescript by gpt-4
+        // This is a port of the Python code from
+        //
+        // Python => Typescript by gpt-4
+        //
+        // https://notebooks.githubusercontent.com/view/ipynb?browser=edge&bypass_fastly=true&color_mode=dark&commit=d67c4181abe9dfd871d382930bb778b7014edc66&device=unknown_device&docs_host=https%3A%2F%2Fdocs.github.com&enc_url=68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f6f70656e61692f6f70656e61692d636f6f6b626f6f6b2f643637633431383161626539646664383731643338323933306262373738623730313465646336362f6578616d706c65732f486f775f746f5f636f756e745f746f6b656e735f776974685f74696b746f6b656e2e6970796e62&logged_in=true&nwo=openai%2Fopenai-cookbook&path=examples%2FHow_to_count_tokens_with_tiktoken.ipynb&platform=mac&repository_id=468576060&repository_type=Repository&version=114#6d8d98eb-e018-4e1f-8c9e-19b152a97aaf
+
         for (const message of messages) {
             num_tokens += tokens_per_message
