+++ src/pricing.ts
@@ -1,3 +1,25 @@
+export type supportModelType =
+    | 'gpt-3.5-turbo'
+    | 'gpt-3.5-turbo-16k'
+    | 'gpt-4'
+    | 'gpt-4-32k'
+    | 'gpt-4-turbo-preview'
+    | 'gpt-3.5-turbo-0301'
+    | 'gpt-3.5-turbo-0613'
+    | 'gpt-3.5-turbo-1106'
+    | 'gpt-3.5-turbo-0125'
+    | 'gpt-3.5-turbo-16k-0613'
+    | 'gpt-4-0314'
+    | 'gpt-4-0613'
+    | 'gpt-4-32k-0314'
+    | 'gpt-4-32k-0613'
+    | 'gpt-4-1106-preview'
+    | 'gpt-4-0125-preview'
+    | 'gpt-4-turbo-2024-04-09'
+    | 'gpt-4-turbo'
+    | 'gpt-4o'
+    | 'gpt-4o-2024-05-13'
+
 /**
  * Pricing
  * @class
